add endpoint set
